requirements.txt
Python
pandas
kafka-python
pandasql